backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
pymongo==4.6.0
motor==3.3.1
joblib==1.3.2
transformers==4.35.2
torch==2.1.1
python-multipart==0.0.6
fastapi-cache2==0.2.1
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0.post1 